add reward bench

+++ eval/alpaca_farm/reward.py
@@ -2,6 +2,7 @@ from transformers import AutoTokenizer, AutoModelForCausalLM, AutoModel, pipelin
 import torch
 import pandas as pd
 import random
+import os
 
 pipe_kwargs = {
         "return_all_scores": True,
@@ -21,35 +22,44 @@ def win_rate(rewards, ref_rewards):
     return sum(winrates) / len(winrates)
     
 def main(args):
-    reward_tokenizer = AutoTokenizer.from_pretrained(args.rm, cache_dir=args.cache_dir)
+    if os.path.exists(args.input_file.replace('.json', '_reward.json')):
+        df = pd.read_json(args.input_file.replace('.json', '_reward.json'), lines=True)
+    else:
+        df = pd.read_json(args.input_file, lines=True)
+        reward_tokenizer = AutoTokenizer.from_pretrained(args.rm, cache_dir=args.cache_dir)
 
-    reward_model = pipeline(
-        "sentiment-analysis",
-        model=args.rm,
-        device_map='auto',
-        tokenizer=reward_tokenizer,
-        model_kwargs={"torch_dtype": torch.float16, "cache_dir": args.cache_dir},
-    )
+        reward_model = pipeline(
+            "sentiment-analysis",
+            model=args.rm,
+            device_map='auto',
+            tokenizer=reward_tokenizer,
+            model_kwargs={"torch_dtype": torch.float16, "cache_dir": args.cache_dir},
+        )
 
-    # load the input file which is a jsonl file
-    df = pd.read_json(args.input_file, lines=True)
-    # turn the df into list of list of dict, which [{'role': 'user', 'content': df['instruction']}, {'role': 'assistant', 'content': df['output']}]
-    dialogues = df.apply(lambda x: [{'role': 'user', 'content': x['instruction']}, {'role': 'assistant', 'content': x['output']}], axis=1).tolist()
-    tokenized_dialogues = [reward_tokenizer.apply_chat_template(text, tokenize=False).replace(reward_tokenizer.bos_token, "") for text in dialogues]
-    with torch.no_grad():
-        rewards = reward_model(tokenized_dialogues, **pipe_kwargs)
-        df['reward'] = [reward[0]['score'] for reward in rewards]
+        # load the input file which is a jsonl file
+        # turn the df into list of list of dict, which [{'role': 'user', 'content': df['instruction']}, {'role': 'assistant', 'content': df['output']}]
+        dialogues = df.apply(lambda x: [{'role': 'user', 'content': x['instruction']}, {'role': 'assistant', 'content': x['output']}], axis=1).tolist()
+        tokenized_dialogues = [reward_tokenizer.apply_chat_template(text, tokenize=False).replace(reward_tokenizer.bos_token, "") for text in dialogues]
+        with torch.no_grad():
+            rewards = reward_model(tokenized_dialogues, **pipe_kwargs)
+            df['reward'] = [reward[0]['score'] for reward in rewards]
 
-    # save the reward to the input file
-    output_file = args.input_file.replace('.json', '_reward.json')
-    df.to_json(output_file, lines=True, orient='records')
+        # save the reward to the input file
+        output_file = args.input_file.replace('.json', '_reward.json')
+        df.to_json(output_file, lines=True, orient='records')
     
     if args.ref_file:
         ref_df = pd.read_json(args.ref_file, lines=True)
         winrate = win_rate(list(df['reward']), list(ref_df['reward']))
         # save as metric json file
-        with open('/'.join(output_file.split('/')[:-1]) + '/metrics.json', 'w') as f:
-            f.write(f'{{"win_rate": {winrate}}}')
+        metric_file = '/'.join(args.input_file.split('/')[:-1]) + '/metrics.json'
+        metrics = {}
+        if os.path.exists(metric_file):
+            with open(metric_file, 'r') as f:
+                metrics = f.read()
+        metrics[os.path.basename(args.ref_file)] = winrate
+        with open(metric_file, 'w') as f:
+            f.write(metrics)
         
 if __name__ == "__main__":
     import argparse
